added a few type casts for compatability purposes

+++ zkdecrypto/z340.cpp
@@ -338,7 +338,7 @@ int ReadNGraphs(const char *filename, int n)
 		
 		//set ngraph
 		if(n==1) unigraphs[index]=percent;
-		else ngraphs[index]=int(10*log(freq));
+		else ngraphs[index]=int(10*log((double)freq));
 	}
 
 	fclose(tgfile);
